Change naming convention for repositories
