+++ core/domain/src/main/java/com/arcticoss/nextplayer/core/domain/SyncMediaUseCase.kt
@@ -1,13 +1,13 @@
 package com.arcticoss.nextplayer.core.domain
 
-import com.arcticoss.nextplayer.core.data.repository.IMediaRepository
+import com.arcticoss.nextplayer.core.data.repository.MediaRepository
 import javax.inject.Inject
 
 class SyncMediaUseCase @Inject constructor(
-    private val mediaRepository: IMediaRepository
+    private val mediaRepository: MediaRepository
 ) {
 
     suspend operator fun invoke() {
-        mediaRepository.syncMedia()
+        mediaRepository.sync()
     }
 }
